refactor: decompose fetch-github-sbom.js into focused modules
Extracts API, parser, slim, writer, and BST modules into
scripts/lib/sbom/. Entry point remains fetch-github-sbom.js.
